the app now requests its initial page from the Lisp process
Because the Electron app requests its initial page from the Lisp
server, we can now generate the correct remote-js context and startup
JS dynamically at launch time, and the generated page correctly
connects to the remote-js websocket when it loads, which in turn gives
us a way for the Lisp to send Javascript expressions to the Electron
app for evaluation. All this now happens automatically, driven by the
Lisp code in presentation-server.lisp.